Extract set- and map-specific code to sub-modules.
